[debug] Widen DebugScriptScopeInfo index and count fields to 32 bits
Three fields in the serialized scope format were too narrow for values
that V8 can actually produce.

1) EncodeAllocInfo() packed the receiver/arguments/function variable slot
   index into 14 signed bits (max 8191). Context slot indices are only
   bounded by the number of variables in the scope, and
   DeclarationScope::AllocateLocals() deliberately allocates the function
   variable last, giving it the highest index in its scope. For

     var g = (function foo() { var v0 .. v8999; eval(""); return foo; });

   the function variable lands in context[9006], which release builds
   silently decoded as -7378. The 'arguments' variable has the same
   problem since it is allocated after all parameters.

2) DebugVariableEntry::slot_index was int16_t, but it stores
   Variable::index(), which for a context- or stack-allocated variable is
   a slot index bounded only by the number of variables in the scope. A
   function declaring 70000 vars alongside a sloppy direct eval puts
   ordinary locals at context[70005]. base::checked_cast CHECK-fails in
   release builds. Parameter indices are stored in the same field but
   stop at Code::kMaxArguments - 1 (65525) and would have fit in 16 bits;
   it is the context and stack slots that force the widening.

3) ScopeRecord::var_count was uint16_t, but a single scope can declare
   far more than 65535 locals.

Widen all three to 32 bits:

- ScopeRecord grows from 16 to 20 bytes (flags -> uint32_t,
  var_count -> int32_t).
- DebugVariableEntry grows from 12 to 14 bytes. It is now packed so that
  slot_index can be int32_t without adding tail padding; the struct is
  only ever read and written as a whole via base::ReadUnalignedValue and
  base::WriteUnalignedValue.
- The packed allocation info word grows from uint16_t to int32_t with a
  30-bit signed index. The manual sign extension is replaced by an
  arithmetic shift, and the range guard is a CHECK rather than a DCHECK
  so that an out-of-range index can never silently corrupt a slot number
  in a release build.

Note that scope record offsets remain only 2-byte aligned, since a
14-byte variable entry means a record's size is not necessarily a
multiple of 4.

flags is widened to 32 bits because the 16 available bits were
exhausted. Once ScopeIterator is fully migrated onto
DebugScriptScopeInfo, some of these flags may turn out to be redundant;
we should revisit then whether the word can shrink back to 16 bits.

TAG=agy
CONV=66e63a01-7fc7-4f67-82be-8f49491976e7

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/8393639
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#109817}

+++ src/debug/debug-scope-info.h
@@ -101,7 +101,7 @@ class V8_EXPORT_PRIVATE DebugScriptScope {
   const uint8_t* payload() const;
   const uint8_t* function_variable_payload() const;
   const uint8_t* variables_payload() const;
-  uint16_t flags() const;
+  uint32_t flags() const;
   int parent_index() const;
 
   // Chained offset calculation methods (private to DebugScriptScope).
